LPS-148050 Modularize Team Members Listing to use custom hooks
